requirements.txt
fastapi
uvicorn
pymupdf
supabase
python-dotenv
openai
python-multipart
azure-ai-inference
azure-identity
